fix: emit correct tag into output tag.json files
The variable `version_json` was used in the loop body. It comes from an unrelated scope before the loop And it contains the last read `version_json`. Instead we need the `version`, that the loop is iterated on. This caused the unrelated data to be emitted into `tag.json`.

This commit renames the loop iterator to `version_json`, overriding the unrelated variable, so that the correct data is used in the loop now.

Python is probably the only language where a local variable declared in a loop can outlive the loop's scope. Perhaps due to be used in a `with` statement? Who knows. Anyways, fantastic language design!