chore(infea): add missing callback invocation for lambda@edge function
This function conditionally rewrites non-gz URLs to gz. The zip files are conditionally excluded. However, in case o zip file the "done" callback was not called at the end of the function, so the response was incorrect.

Here I add the callback invocation at the end of the function, so that requests for zip files are correctly replied to.