Cleanup datasets
Here I cleanup datasets:
-  If I don't recognize a dataset or if I know that it is incorrect, I removed it.
- I give friendly names to the remaining datasets, mostly according to what's written in their reference fasta file.
- I try to make directory structure and naming a little more consistent (it is now reflected in the dataset `path` which is used as an identifier of the dataset in many places exposed to the end user).

This is mostly a "cosmetic" change, needed for testing usability of the current Nextclade Web UI and for making it prettier.

I have little knowledge in most of these pathogens, so this PR can be rolled back and the reorganization can be done properly by a scientists later.